server_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
absl-py=1.4.0=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
albumentations=1.3.0=pypi_0
astor=0.8.1=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.2.0=pypi_0
blas=1.0=mkl
blosc=1.21.0=h4ff587b_1
bottleneck=1.3.5=py310ha9d4c09_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.12.7=ha878542_0
cachetools=5.3.0=pypi_0
captum=0.6.0=0
certifi=2022.12.7=pyhd8ed1ab_0
charset-normalizer=3.1.0=pypi_0
cloudpickle=2.2.1=pypi_0
cmake=3.26.1=pypi_0
colorama=0.4.6=pypi_0
contextlib2=21.6.0=pypi_0
contourpy=1.0.5=py310hdb19cb5_0
cudatoolkit=11.3.1=h2bc3f7f_2
curl=7.84.0=h5eee18b_0
cycler=0.11.0=pyhd3eb1b0_0
dnspython=2.3.0=pypi_0
expat=2.4.4=h295c915_0
filelock=3.10.7=pypi_0
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=pypi_0
fsspec=2023.3.0=pypi_0
future=0.18.3=pypi_0
gettext=0.21.0=hf68c758_0
giflib=5.2.1=h5eee18b_1
git=2.34.1=pl5262hc120c5b_0
git-lfs=3.3.0=ha770c72_0
google-auth=2.16.3=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.53.0=pypi_0
hdf5=1.12.1=h70be1eb_2
hyperopt=0.1.2=pypi_0
icu=58.2=he6710b0_3
idna=3.4=pypi_0
imageio=2.27.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=h7f8727e_0
json-tricks=3.16.1=pypi_0
kiwisolver=1.4.4=py310h6a678d5_0
krb5=1.19.2=hac12032_0
lazy-loader=0.2=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcurl=7.84.0=h91b91d3_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libffi=3.4.2=h6a678d5_6
libgcc-ng=12.2.0=h65d4601_19
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libnghttp2=1.46.0=hce63b2e_0
libpng=1.6.37=hbc83047_0
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.0=hecacb30_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxml2=2.9.14=h74e7548_0
lightning-utilities=0.8.0=pypi_0
lit=16.0.0=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.3=h295c915_1
lzo=2.10=h7b6447c_2
markdown=3.4.3=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib-base=3.6.2=py310h945d387_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
mock=4.0.3=pyhd3eb1b0_0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
munkres=1.1.4=py_0
ncurses=6.3=h5eee18b_3
networkx=3.0=pypi_0
nni=2.9=pypi_0
numexpr=2.8.3=py310hcea2de6_0
numpy=1.23.1=py310h1794996_0
numpy-base=1.23.1=py310hcba007f_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.7.0.72=pypi_0
opencv-python-headless=4.7.0.72=pypi_0
openssl=1.1.1t=h0b41bf4_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.3=py310h6a678d5_0
patchify=0.2.3=pypi_0
pcre2=10.37=he7ceb23_1
perl=5.26.2=h14c3975_0
pillow=9.3.0=py310hace64e9_1
pip=22.3.1=py310h06a4308_0
prettytable=3.6.0=pypi_0
protobuf=4.22.1=pypi_0
psutil=5.9.4=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pymongo=4.3.3=pypi_0
pyparsing=3.0.4=pyhd3eb1b0_0
pytables=3.7.0=py310h9da3b7f_0
python=3.10.8=h7a1cb2a_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pythonwebhdfs=0.2.3=pypi_0
pytorch-lightning=2.0.0=pypi_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py310h06a4308_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
qudida=0.0.4=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
responses=0.23.1=pypi_0
rsa=4.9=pypi_0
schema=0.7.5=pypi_0
scikit-image=0.20.0=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
setuptools=65.6.3=py310h06a4308_0
simplejson=3.18.4=pypi_0
six=1.16.0=pyhd3eb1b0_1
split-folders=0.5.1=pypi_0
sqlite=3.40.1=h5082296_0
sympy=1.11.1=pypi_0
tensorboard=2.12.0=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2023.3.21=pypi_0
tk=8.6.12=h1ccaba5_0
torch=2.0.0=pypi_0
torchmetrics=0.11.4=pypi_0
torchvision=0.15.1=pypi_0
tqdm=4.65.0=pypi_0
tree=2.1.0=h0b41bf4_0
triton=2.0.0=pypi_0
typeguard=2.13.3=pypi_0
types-pyyaml=6.0.12.9=pypi_0
typing_extensions=4.4.0=py310h06a4308_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.15=pypi_0
wcwidth=0.2.6=pypi_0
websockets=10.4=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.10=h5eee18b_1
yarl=1.8.2=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
